refactor: split load test report Go implementation
Separate the exploratory Go report implementation into focused files for argument parsing, Prometheus access, query loading, report construction, and rendering. This keeps the comparison PR easier to review while preserving the stable shell entrypoint.

Use typed query and Prometheus response structs with encoding/json for JSON handling. Replace the shell-style curl option parser with explicit bearer and basic-auth flags because the report only needs those authentication modes.
